examples/ping: more libc::poll() fixes

+++ examples/ping/src/main.rs
@@ -237,7 +237,7 @@ pub fn listen_multiplex<S: ?Sized + AsRef<str>, H: ::ConnectionHandler + Send + 
         let r = unsafe {
             libc::poll(
                 fds.as_mut_ptr(),
-                fds.len().into(),
+                (fds.len() as u32).into(),
                 (accept_timeout * 1000) as i32,
             )
         };
